+++ astro/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @bs-core/astro
 
+## 1.8.20
+
+### Patch Changes
+
+- Updated package.json to export types in a better way
+- Updated dependencies
+- Updated dependencies
+  - @bs-core/shell@1.21.11
+
 ## 1.8.19
 
 ### Patch Changes